requirements.txt
openai
python-docx
python-dotenv
flask